move footer to index
